Make refactoring in tests
