+++ contracts/avida_example/Cargo.toml
@@ -3,18 +3,30 @@ name = "avida-example"
 version = "0.1.0"
 edition = "2021"
 authors = ["andrew <andrew@nymlab.it>"]
+resolver = "2"
+
+exclude = [
+  "contract.wasm",
+  "hash.txt",
+]
 
 [lib]
 crate-type = ["cdylib", "rlib"]
 doctest = false
 
+[features]
+backtraces = ["cosmwasm-std/backtraces"]
+library = []
+migration = []
+mt = ["sylvia/mt"]
+
 [dependencies]
-avida-sdjwt-verifier = { path = "../sdjwt-verifier/"}
+avida-sdjwt-verifier = { path = "../sdjwt-verifier/", features = ["library"]}
 avida-common = { path = "../../packages/common/"}
 cosmwasm-schema = { workspace = true }
 schemars = { workspace = true }
 cosmwasm-std = { workspace = true }
-sylvia = { workspace = true, features = ["mt"]}
+sylvia = { workspace = true}
 serde = { workspace = true }
 cw-storage-plus.workspace = true
 cw-utils = { workspace = true }
@@ -25,4 +37,5 @@ serde_json = "1.0.116"
 josekit = "0.8.6"
 sd-jwt-rs = "0.7.0"
 jsonwebtoken = { version="9.3.0", features=["use_pem"]}
-cw-multi-test = { version = "0.20"}
+sylvia = { workspace = true, features = ["mt"]}
+avida-test-utils = { path = "../../packages/avida_test_utils/", features = ["sdjwt"]}
