+++ contracts/avida_example/schema/raw/instantiate.json
@@ -1,0 +1,13 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "InstantiateMsg",
+  "type": "object",
+  "required": [
+    "verifier"
+  ],
+  "properties": {
+    "verifier": {
+      "type": "string"
+    }
+  }
+}
