+++ contracts/avida_example/schema/raw/response_to_get_verifier_address.json
@@ -1,0 +1,14 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "GetVerifierResponse",
+  "type": "object",
+  "required": [
+    "verifier"
+  ],
+  "properties": {
+    "verifier": {
+      "type": "string"
+    }
+  },
+  "additionalProperties": false
+}
