+++ contracts/avida_example/src/tests/mod.rs
@@ -1,4 +1,5 @@
 pub mod flow_test;
 pub mod query_test;
 pub mod exec_test;
+
 pub mod fixtures;
