+++ contracts/sdjwt-verifier/src/lib.rs
@@ -2,3 +2,6 @@ pub mod contract;
 pub mod errors;
 pub mod types;
 mod verifier;
+
+#[cfg(test)]
+mod tests;
