+++ contracts/sdjwt-verifier/src/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod verifier_test;
+mod fixtures;
