+++ packages/avida_test_utils/src/lib.rs
@@ -1,4 +1,2 @@
-pub mod util;
-
-#[cfg(test)]
-mod sdjwt;
+#[cfg(feature = "sdjwt")]
+pub mod sdjwt;
