+++ packages/avida_test_utils/src/sdjwt/fixtures.rs
@@ -7,17 +7,10 @@ use std::{fs, path::PathBuf};
 
 use cosmwasm_std::Binary;
 
-use sylvia::multitest::{App, Proxy};
-
 use avida_common::types::{
     InputRoutesRequirements, RouteVerificationRequirements, VerificationSource,
 };
-use avida_sdjwt_verifier::{
-    contract::sv::mt::CodeId,
-    contract::*,
-    types::{Criterion, InitRegistration, MathsOperator, PresentationReq},
-};
-use cw_multi_test::App as MtApp;
+use avida_sdjwt_verifier::types::{Criterion, MathsOperator, PresentationReq};
 use josekit::{self};
 
 /// Test constants
@@ -232,38 +225,4 @@ pub fn get_input_route_requirement(
             route_verification_requirements_type,
         ),
     }
-}
-
-/// Is used to instantiate verifier contract with some predefined parameters
-pub fn instantiate_verifier_contract<'a>(
-    app: &'a App<MtApp>,
-    route_verification_requirements_type: RouteVerificationRequirementsType,
-) -> (
-    Proxy<'a, MtApp, SdjwtVerifier<'a>>,
-    RouteVerificationRequirements,
-) {
-    let fx_route_verification_req =
-        get_route_verification_requirement(route_verification_requirements_type);
-    let code_id = CodeId::store_code(app);
-
-    // String, // Admin
-    // String, // App Addr
-    // Vec<(RouteId, RouteVerificationRequirements)>,
-    let init_registrations = vec![InitRegistration {
-        app_admin: FIRST_CALLER_APP_ADDR.to_string(),
-        app_addr: FIRST_CALLER_APP_ADDR.to_string(),
-        routes: vec![InputRoutesRequirements {
-            route_id: FIRST_ROUTE_ID,
-            requirements: fx_route_verification_req.clone(),
-        }],
-    }];
-
-    (
-        code_id
-            .instantiate(MAX_PRESENTATION_LEN, init_registrations)
-            .with_label(VERIFIER_CONTRACT_LABEL)
-            .call(OWNER_ADDR)
-            .unwrap(),
-        fx_route_verification_req,
-    )
 }
