--- packages/avida_test_utils/src/util.rs
@@ -1,1 +1,0 @@
-
