+++ scripts/schemas.sh
@@ -1,5 +1,5 @@
 
-for dir in $PWD/contracts/sdjwt-verifier/; do
+for dir in $PWD/contracts/sdjwt-verifier/ $PWD/contracts/avida_example; do
  cd $dir
  cargo run
  cd -
